requirements.txt
streamlit
pandas
psycopg2-binary
sqlalchemy
matplotlib
pyarrow